♻️ refactor: api.ts => enum.ts로 파일명 변경
